pyproject.toml
[project]
name = "orchard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "uvicorn>=0.35.0",
    "jsonschema>=4.23.0",
    "llama-index-core>=0.11.23",
    "llama-index-readers-github>=0.2.0",
    "llama-index-readers-web>=0.1.0",
    "llama-index-readers-google>=0.2.0",
    "llama-index-embeddings-huggingface>=0.3.1",
    "llama-index-vector-stores-chroma>=0.3.0",
    "gitpython>=3.1.43",
    "beautifulsoup4>=4.12.3",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.2",
]
